saving and  routine addendum

+++ header_files/lcd_definitions.h
@@ -199,7 +199,7 @@ void notSavedScreen(void);
 
 
 
-uint32_t flashData2LCD(float x,uint32_t n);
+uint32_t flashData2LCD(uint32_t x,uint32_t n);
 uint32_t screenData2flash(uint32_t* data);
 
 #endif
